clean up duplicated timing logic
